refactor(http): 패키지 구조 변경
request, response 패키지를 http 패키지 하위로 이동
