+++ src/main/java/webserver/controller/Controller.java
@@ -1,7 +1,7 @@
 package webserver.controller;
 
-import webserver.request.HttpRequest;
-import webserver.response.HttpResponse;
+import webserver.http.request.HttpRequest;
+import webserver.http.response.HttpResponse;
 
 public interface Controller {
     HttpResponse service(HttpRequest request);
