+++ src/main/java/webserver/http/HttpCookie.java
@@ -1,0 +1,21 @@
+package webserver.http;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public class HttpCookie {
+
+    private final Map<String, String> cookies;
+
+    public HttpCookie(Map<String, String> cookies) {
+        this.cookies = cookies;
+    }
+
+    public HttpCookie() {
+        this(new HashMap<>());
+    }
+
+    public void setCookie(String key, String value) {
+
+    }
+}
