+++ src/main/java/webserver/http/request/HttpRequestFactory.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 import utils.IOUtils;
 import utils.parser.DataFormat;
