+++ src/main/java/webserver/http/request/HttpRequestHeader.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 import java.util.HashMap;
 import java.util.List;
