+++ src/main/java/webserver/http/request/HttpRequestMethod.java
@@ -1,4 +1,4 @@
-package webserver.request;
+package webserver.http.request;
 
 public enum HttpRequestMethod {
     GET,
