+++ src/main/java/webserver/http/response/StatusLine.java
@@ -1,4 +1,4 @@
-package webserver.response;
+package webserver.http.response;
 
 public class StatusLine {
